minor fixes

+++ MatchmakerServer/NetworkLibrary/Http/GameUnit.cs
@@ -5,7 +5,7 @@ namespace NetworkLibrary.NetworkLibrary.Http
     [ZeroFormattable]
     public class GameUnit
     {
-        [Index(0)] public virtual int TemporaryId { get; set; }
+        [Index(0)] public virtual ushort TemporaryId { get; set; }
         [Index(1)] public virtual bool IsBot { get; set; }
         [Index(2)] public virtual string PrefabName { get; set; }
         /// <summary>
